% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{BackgroundWorker}
\alias{BackgroundWorker}
\alias{$.BackgroundWorker}
\title{Represents a background running R script.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
BackgroundWorker
}
\description{
Represents a background running R script.
}
\keyword{datasets}
